Use different docker-watcher not launched through docker-compose
Without user-namespaces the docker-watcher can be ran in a docker.  As
we use user-namespaces in the development setup of mu-semtech, that is
not an option during development.  In order to use the same setup,
updates happened to this component.

The whole setup can be launched by calling the new launch.sh script.